Prompt to delete local branches which no longer exist on the remote.
When checking out a branch locally for review, it is common that it will
later get updated on the remote via a force-push (to clean up fixups etc),
merged and then deleted. When this happens, the local branch does not
match what was merged into master exactly, so the previous logic fails
to delete it.

Provide a way to delete such branches, but prompt because it is an
unsafe operation.

This is based on comments in https://stackoverflow.com/questions/15661853