Set `LANG` env var in Alacritty config
This variable was not being set by Alacritty. As a result when running
applications in tox, which only allows `LANG` but not `LC_*` env vars
from the parent process through to child processes, the default encoding
for reading/writing text files was defaulting to US-ASCII rather than
UTF-8. This caused encoding errors if files contained non-ASCII
characters.