Use a more reliable method of determining the HEAD branch of the remote
After the main branch was renamed from "master" to "main" in various
Hypothesis projects, I found that the `.git/refs/remotes/origin/HEAD`
file still referenced the "master" branch. Using `git remote show
origin` however, which actually queries the remote, returned the updated
branch in the `HEAD branch: <name>` line. The downside of this approach
is that it involves querying the remote and thus doesn't work if
offline.